refactor: 💡 Refactor auth packages in api
