+++ apps/api/src/main/java/dk/treecreate/api/authentication/AuthController.java
@@ -1,16 +1,16 @@
-package dk.treecreate.api.controllers;
+package dk.treecreate.api.authentication;
 
-import dk.treecreate.api.models.ERole;
-import dk.treecreate.api.models.Role;
-import dk.treecreate.api.models.User;
-import dk.treecreate.api.payload.request.LoginRequest;
-import dk.treecreate.api.payload.request.SignupRequest;
-import dk.treecreate.api.payload.response.JwtResponse;
-import dk.treecreate.api.payload.response.MessageResponse;
-import dk.treecreate.api.repository.RoleRepository;
-import dk.treecreate.api.repository.UserRepository;
-import dk.treecreate.api.security.jwt.JwtUtils;
-import dk.treecreate.api.security.services.UserDetailsImpl;
+import dk.treecreate.api.authentication.models.ERole;
+import dk.treecreate.api.authentication.models.Role;
+import dk.treecreate.api.authentication.models.User;
+import dk.treecreate.api.authentication.dto.request.LoginRequest;
+import dk.treecreate.api.authentication.dto.request.SignupRequest;
+import dk.treecreate.api.authentication.dto.response.JwtResponse;
+import dk.treecreate.api.authentication.dto.response.MessageResponse;
+import dk.treecreate.api.authentication.repository.RoleRepository;
+import dk.treecreate.api.authentication.repository.UserRepository;
+import dk.treecreate.api.authentication.jwt.JwtUtils;
+import dk.treecreate.api.authentication.services.UserDetailsImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
