+++ apps/api/src/main/java/dk/treecreate/api/authentication/TestController.java
@@ -1,4 +1,4 @@
-package dk.treecreate.api.controllers;
+package dk.treecreate.api.authentication;
 
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.CrossOrigin;
