+++ apps/api/src/main/java/dk/treecreate/api/authentication/jwt/AuthTokenFilter.java
