+++ apps/api/src/test/java/dk/treecreate/api/healthcheck/HealthcheckTests.java
@@ -1,8 +1,8 @@
 package dk.treecreate.api.healthcheck;
 
-import dk.treecreate.api.security.jwt.AuthEntryPointJwt;
-import dk.treecreate.api.security.jwt.JwtUtils;
-import dk.treecreate.api.security.services.UserDetailsServiceImpl;
+import dk.treecreate.api.authentication.jwt.AuthEntryPointJwt;
+import dk.treecreate.api.authentication.jwt.JwtUtils;
+import dk.treecreate.api.authentication.services.UserDetailsServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
