+++ apps/api/src/test/java/dk/treecreate/api/mail/MailControllerTest.java
@@ -3,9 +3,9 @@ package dk.treecreate.api.mail;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import dk.treecreate.api.mail.dto.ResetPasswordDto;
 import dk.treecreate.api.mail.dto.SignupDto;
-import dk.treecreate.api.security.jwt.AuthEntryPointJwt;
-import dk.treecreate.api.security.jwt.JwtUtils;
-import dk.treecreate.api.security.services.UserDetailsServiceImpl;
+import dk.treecreate.api.authentication.jwt.AuthEntryPointJwt;
+import dk.treecreate.api.authentication.jwt.JwtUtils;
+import dk.treecreate.api.authentication.services.UserDetailsServiceImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
